web/requirements.txt
flask==3.0.3
pandas==2.2.2
requests==2.32.3
Flask-Caching==2.3.0
redis==5.0.8
python-dotenv
